requirement.txt
llama-index==0.14.13
llama-index-embeddings-huggingface
fastmcp==2.14.5
pydantic-ai
pandas
requests
openai
python-dotenv